Project restructuring and added registration page
The goal of restructuring is to organize the project in:
- pages
- components
- services
- pipes
- etc

Cards on 'Promociones' tab located inside a responsive grid for better appearance.

Added registration and promo detail page

Also added temporal menu on toolbar, to allow us to reach registration page and other pages that could be in development but their location is not yet defined.